Fixed Drawing clipping issue with DPI change and added closing of window after drawing noise map

+++ xper-train/xper-allen/src/org/xper/allen/util/DPIUtil.java
@@ -6,11 +6,11 @@ import org.xper.drawing.renderer.AbstractRenderer;
 /**
  * Job is to take the monkey screen's DPI (dots per inch), and maximum stimulus size in degrees that is expected,
  *  in order to calculate
- * the bare minimum pixel resolution that would achieve at least that DPI at the maximum stimulus size. 
- * 
+ * the bare minimum pixel resolution that would achieve at least that DPI at the maximum stimulus size.
+ *
  * Another functionality is to take the DPI of the monitor that the images are being generated on
- * and use this to calculate the size of the openGL window in mm so the images can be generated 
- * properly. 
+ * and use this to calculate the size of the openGL window in mm so the images can be generated
+ * properly.
  * @author Allen Chen
  *
  */
@@ -23,22 +23,22 @@ public class DPIUtil {
 	double maxImageDimensionDegrees;
 	@Dependency
 	AbstractRenderer renderer;
-	
+
 	public double maxDimMm;
 	public int calculateMinResolution(){
 		maxDimMm = renderer.deg2mm(maxImageDimensionDegrees);
 		double maxDimInches = maxDimMm / 25.4;
 		int minPixels = (int) Math.round(monkeyDPI*maxDimInches);
-		
+
 		return minPixels+1;
 	}
 
 	public double calculateMmForRenderer() {
 		int pixels = calculateMinResolution();
 		//System.out.println(pixels * (1/generatorDPI) * 25.4);
-		return pixels * (1/generatorDPI) * 25.4;
+		return (double) pixels * (1.0/generatorDPI) * 25.4;
 	}
-	
+
 	public double getDpi() {
 		return monkeyDPI;
 	}
